+++ changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.19] - 2026-09-06
+
 ### Fixed
 
 - Scrolling the read-only preview quickly and stopping no longer leaves most of